refactor(blog-page): cleaned up blog page code a bit
